chore(hog-charts): trim TimeSeriesLineChart api + reorg utils
Three internal cleanups of lib/hog-charts that the recently-shipped
trends migration revealed as dead surface area:

- Trim TimeSeriesLineChartConfig: drop config.anomalies (TrendsAlertOverlays
  renders AnomalyPointsLayer as a child overlay), config.inProgress (use
  per-series Series.stroke.partial instead), and the auto-X-axis-formatter
  path (xAxis.timezone / interval / allDays). xAxis keeps tickFormatter
  + hide; callers that want date-axis ticks pass createXAxisTickCallback
  in directly. Deletes utils/in-progress.ts.

- Move AnomalyPointsLayer up to lib/hog-charts/overlays/ alongside
  ReferenceLine / ValueLabels (it's a generic primitive, not
  TimeSeriesLineChart-specific).

- Reorganise lib/hog-charts/utils/: dates, y-formatters, goal-lines,
  and use-axis-formatters now live at the package level (cross-cutting,
  not TimeSeriesLineChart-specific). Extract applyComparisonDimming +
  dimHex into utils/comparison-dimming.ts. derived-series.ts keeps the
  CI / moving-avg / trend-line builders and imports dimHex from there.
