+++ frontend/src/lib/hog-charts/charts/TimeSeriesLineChart/TimeSeriesLineChart.stories.tsx
@@ -1,7 +1,7 @@
 import { Meta, StoryObj } from '@storybook/react'
 
-import { DEFAULT_Y_AXIS_ID, TimeSeriesLineChart } from 'lib/hog-charts'
-import type { AnomalyMarker, Series, TimeInterval, YAxisConfig } from 'lib/hog-charts'
+import { createXAxisTickCallback, TimeSeriesLineChart } from 'lib/hog-charts'
+import type { Series, TimeInterval, YAxisConfig } from 'lib/hog-charts'
 import { ciRanges } from 'lib/statistics'
 
 import { Stage, useReactiveTheme } from '../../story-helpers'
@@ -88,6 +88,7 @@ interface DateAxisCellProps {
 
 function DateAxisCell({ title, labels, series, interval, timezone }: DateAxisCellProps): JSX.Element {
     const theme = useReactiveTheme()
+    const tickFormatter = createXAxisTickCallback({ timezone, interval, allDays: labels })
     return (
         // eslint-disable-next-line react/forbid-dom-props
         <div style={{ display: 'flex', flexDirection: 'column', gap: 4 }}>
@@ -98,7 +99,7 @@ function DateAxisCell({ title, labels, series, interval, timezone }: DateAxisCel
                     labels={labels}
                     theme={theme}
                     config={{
-                        xAxis: { timezone, interval },
+                        xAxis: { tickFormatter },
                         yAxis: { showGrid: true },
                     }}
                 />
@@ -164,29 +165,6 @@ export const YAxisFormats: Story = {
             <YFormatCell title="duration_ms" series={DURATION_MS_SERIES} config={{ format: 'duration_ms' }} />
         </div>
     ),
-}
-
-export const Anomalies: Story = {
-    render: () => {
-        const theme = useReactiveTheme()
-        const series: Series[] = [{ key: 'visits', label: 'Visits', data: [20, 35, 28, 60, 45, 70, 52] }]
-        // Severity drives color — low (amber), medium (orange), high (red).
-        const anomalies: AnomalyMarker[] = [
-            { dataIndex: 1, value: 35, color: '#f59e0b', yAxisId: DEFAULT_Y_AXIS_ID },
-            { dataIndex: 3, value: 60, color: '#f97316', yAxisId: DEFAULT_Y_AXIS_ID },
-            { dataIndex: 5, value: 70, color: '#ef4444', yAxisId: DEFAULT_Y_AXIS_ID },
-        ]
-        return (
-            <Stage>
-                <TimeSeriesLineChart
-                    series={series}
-                    labels={DAYS}
-                    theme={theme}
-                    config={{ yAxis: { showGrid: true }, anomalies }}
-                />
-            </Stage>
-        )
-    },
 }
 
 const DERIVED_SERIES: Series[] = [
