+++ frontend/src/lib/hog-charts/utils/dates.test.ts
