+++ frontend/src/lib/hog-charts/utils/dates.ts
