+++ frontend/src/lib/hog-charts/utils/y-formatters.test.ts
