+++ frontend/src/lib/hog-charts/utils/y-formatters.ts
